+++ UILabel_Typography_Extensions/Extensions/UIKit+Extensions.swift
@@ -46,8 +46,8 @@ extension UIView {
 		with {
 			$0.layer.borderWidth = 1
 			$0.layer.cornerRadius = 2
-			$0.layer.borderColor = UIColor.red.withAlphaComponent(0.3).cgColor
-			// $0.backgroundColor = UIColor.red.withAlphaComponent(0.1)
+			$0.layer.borderColor = UIColor.red.withAlphaComponent(0.6).cgColor
+			$0.backgroundColor = UIColor.red.withAlphaComponent(0.2)
 		}
 	}